feat(client): create variable on NotebookCell, Layers, and SelectedMenu
