+++ packages/client/src/assets/img/Database.tsx
@@ -1,6 +1,10 @@
 import React from 'react';
 
-export const Database = () => {
+interface IconProps {
+    color?: string;
+}
+export const Database = (props: IconProps) => {
+    const { color = '#EBEEFE' } = props;
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -14,7 +18,7 @@ export const Database = () => {
                     fillRule="evenodd"
                     clipRule="evenodd"
                     d="M12 3C7.58 3 4 4.79 4 7V17C4 19.21 7.59 21 12 21C16.41 21 20 19.21 20 17V7C20 4.79 16.42 3 12 3ZM18 17C18 17.5 15.87 19 12 19C8.13 19 6 17.5 6 17V14.77C7.61 15.55 9.72 16 12 16C14.28 16 16.39 15.55 18 14.77V17ZM18 12.45C16.7 13.4 14.42 14 12 14C9.58 14 7.3 13.4 6 12.45V9.64C7.47 10.47 9.61 11 12 11C14.39 11 16.53 10.47 18 9.64V12.45ZM12 9C8.13 9 6 7.5 6 7C6 6.5 8.13 5 12 5C15.87 5 18 6.5 18 7C18 7.5 15.87 9 12 9Z"
-                    fill="#EBEEFE"
+                    fill={color}
                 />
             </g>
             <defs>
