+++ packages/client/src/components/blocks-workspace/BlocksWorkspace.tsx
@@ -3,12 +3,8 @@ import { observer } from 'mobx-react-lite';
 import { useNotification, styled, Typography, Stack } from '@semoss/ui';
 
 import { runPixel } from '@/api';
-import {
-    STATE_STORE_CURRENT_VERSION,
-    SerializedState,
-    StateStore,
-    WorkspaceStore,
-} from '@/stores';
+import { SerializedState, StateStore, WorkspaceStore } from '@/stores';
+import { STATE_STORE_CURRENT_VERSION } from '@/stores/state/MigrationManager';
 import { DefaultCells } from '@/components/cell-defaults';
 import { DefaultBlocks } from '@/components/block-defaults';
 import { Blocks } from '@/components/blocks';
